+++ docs/changelog.md
@@ -2,6 +2,9 @@
 Changelog
 ============
 
+* UNRELEASED
+  - Save volume levels
+
 * Release 0.7.9 - 2025-03-25
   - Fix wallpaper fade in on launch
   - Improve detection of window movements in desktop switcher